Makes media uploadable after the claim is created
Since it's not possible to know when the claim becomes complete without
knowing which media files have been already uploaded in such conditions,
we're asking for the array of such filenames and then check whether
that's all of them uploaded. In such case the claim is deemed complete.

This way client can also be aware, which files have successfully reached
the target and which need to be reuploaded in case of network or other
kind of failure.

- New API resource to upload claim media
- Claims now have a `media_filenames` array field that should be used to
  contain original filenames of the media to be uploaded
- Same goes for the media model, which now posesses an
  `original_filename` field since the uploaded filename will now become
  a UUID to avoid guessing
- Renamed `images` to `media` for generalisation (we might have video
  files in the future)
- Added some more claim statuses for completion tracking and with
  moderation in mind
- Claim cancel is now logged as a state
- Removed Ukrainian labels from the models for consistency, we'll do a
  proper i18n later if needed
- Upgraded DRF to 3.4.0 and DRS to 0.3.8

Closes #19.